Add files via upload
Summary of the work done of to each of the files:

main - nothing changed except defined the hash function within the scope of this file. we can hash each word as we take them in from the string stream, then perform inserts to the b tree using the hash as a parameter.

btree.h - added a boolean to the BTNode class that will tell us if the node is a leaf or not. added a currentKeys int variable to BTNode that will keep track of how many elements are in each node. declared all methods as methods of the BTree class. changed key var to minDeg (aka minimum degree) since much of the arithmetic in the code centers around this value, see comments for explanation. certain methods like splitChild and insertNonFull can be within the scope of BTNode, but passing in nodes as parameters should work fine.

btree.cpp - implemented parameterized constructors for both BTNode and BTree. implemented search method. implemented insert method. implemented splitChild and insertNonFull method since they are needed for insert. as of right now, i don't think the code accounts for duplicate keys, this is likely the main thing we need to figure out before the tree will fully work.