Code closer to compiling
Worked on polishing up the code so it will compile. Still does not compile, but is much closer. The main issue I am having right now is that the compiler says the member functions of BTree cannot access the private data members of BTNode, even though declaring BTree as a friend class of BTNode should take care of this issue. Also, I added the code that will read each word from the text file and insert it into the tree if the word has not already been inserted.