Creates tree but still work in progress.
Eliminated the seg fault by changing around indexes. Index in part of the code was off by one. As far as i can tell, the tree creates itself fine but there is something weird going on with the repeated words. Sometimes count is incremented and other times it is not, in which case the repeated word is added to the tree. Going to try to figure that out over the next few days.