Fixed error with search function
Off by one index error was causing certain parts of the tree to not be searched, resulting in the inconsistencies seen with the count variable and duplicate words. Tree should be fully functional now, just need to make the dot visualization to ensure that it looks how it should and doesn't break any rules of a b tree.